+++ mordant/src/commonTest/kotlin/com/github/ajalt/mordant/widgets/PaddingTest.kt
@@ -7,64 +7,110 @@ class PaddingTest : RenderingTest() {
     private val w = Text("x")
 
     @Test
-    fun oneValue() = checkRender(w.withPadding(1), """
+    fun oneValue() = checkRender(
+        w.withPadding(1), """
     ░░
     ░ x ░
     ░░
-    """)
+    """
+    )
 
     @Test
-    fun twoValues() = checkRender(w.withPadding(1, 2), """
+    fun twoValues() = checkRender(
+        w.withPadding(1, 2), """
     ░░
     ░  x  ░
     ░░
-    """)
+    """
+    )
 
     @Test
-    fun threeValues() = checkRender(w.withPadding(1, 2, 3), """
+    fun threeValues() = checkRender(
+        w.withPadding(1, 2, 3), """
     ░░
     ░  x  ░
     ░░
     ░░
     ░░
-    """)
+    """
+    )
 
     @Test
-    fun fourValues() = checkRender(w.withPadding(1, 2, 3, 4), """
+    fun fourValues() = checkRender(
+        w.withPadding(1, 2, 3, 4), """
     ░░
     ░    x  ░
     ░░
     ░░
     ░░
-    """)
+    """
+    )
 
     @Test
-    fun vertical() = checkRender(w.withVerticalPadding(1), """
+    fun vertical() = checkRender(
+        w.withVerticalPadding(1), """
     ░░
     ░x░
     ░░
-    """)
+    """
+    )
 
     @Test
-    fun horizontal() = checkRender(w.withHorizontalPadding(1), """
+    fun horizontal() = checkRender(
+        w.withHorizontalPadding(1), """
     ░ x ░
-    """)
+    """
+    )
 
     @Test
-    fun padEmpty() = checkRender(Text("x\n\ny").withPadding(1, padEmptyLines = true), """
+    fun padEmpty() = checkRender(
+        Text("x\n\ny").withPadding(1, padEmptyLines = true), """
     ░░
     ░ x ░
     ░  ░
     ░ y ░
     ░░
-    """)
+    """
+    )
 
     @Test
-    fun noPadEmpty() = checkRender(Text("x\n\ny").withPadding(1, padEmptyLines = false), """
+    fun noPadEmpty() = checkRender(
+        Text("x\n\ny").withPadding(1, padEmptyLines = false), """
     ░░
     ░ x ░
     ░░
     ░ y ░
     ░░
-    """)
+    """
+    )
+
+    @Test
+    fun withTopPadding() = checkRender(
+        w.withTopPadding(1), """
+    ░░
+    ░x░
+    """
+    )
+
+    @Test
+    fun withRightPadding() = checkRender(
+        w.withRightPadding(1), """
+    ░x ░
+    """
+    )
+
+    @Test
+    fun withBottomPadding() = checkRender(
+        w.withBottomPadding(1), """
+    ░x░
+    ░░
+    """
+    )
+
+    @Test
+    fun withLeftPadding() = checkRender(
+        w.withLeftPadding(1), """
+    ░ x░
+    """
+    )
 }