+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Implemented `hideInput` for prompts on native targets [(#63)](https://github.com/ajalt/mordant/issues/63)
 - Improve cell-width calculation for emoji sequences like skin tone modifiers. [(#64)](https://github.com/ajalt/mordant/issues/64)
+- `withTopPadding`, `withRightPadding`, `withBottomPadding`, `withLeftPadding` extensions for adding padding to single side of a widget
 
 ## 2.0.0-beta7
 ### Added